refactor: turn RunParser into a proper TS pattern
